pyproject.toml
[project]
name = "market-bot"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "alpaca-py>=0.43.4",
    "discord-py>=2.7.1",
    "fastapi>=0.128.8",
    "psycopg[binary]>=3.2.13",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.39.0",
]
